feat: project renamed
