+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "saao-deployment-server"
+name = "remote-command-server"
 version = "0.1.0"
 description = ""
 authors = ["Christian Hettlage <hettlage@saao.ac.za>"]
@@ -9,6 +9,7 @@ python = "^3.7"
 fastapi = {extras = ["all"], version = "^0.63.0"}
 uvicorn = {extras = ["standard"], version = "^0.13.3"}
 SQLAlchemy = "^1.3.22"
+click = "^7.1.2"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.1"
@@ -18,6 +19,9 @@ tox = "^3.21.2"
 isort = "^5.7.0"
 mypy = "^0.790"
 pytest-cov = "^2.11.0"
+
+[tool.poetry.scripts]
+rcs = 'remote_command_server.cli:cli'
 
 [tool.isort]
 multi_line_output = 3
